healthcare_backend/requirements_production.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2

# Production dependencies for Render
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
